Don't delete the save until you die
This means that if the game crashes, you don't lose your save.  This also
means that it's possible to cheat with signals and power cuts.  I don't
see this as a big problem, since you can already cheat with cp (and ptrace,
which uses the same access checks as non-terminal signals).